feat: add more test
