+++ extensions/iceworks-refactor/src/test/expects/TodoPage.tsx
@@ -1,15 +1,8 @@
 import React, { Component } from 'react';
-import Todo, { Todo2 } from '../../components/Todo';
-
-const a = 12;
-
 export default class TodoComponent extends Component {
   text1: number;
-
   name2: string;
-
   name3: string;
-
   name4 = 'a';
 
   constructor(props) {
@@ -24,12 +17,6 @@ export default class TodoComponent extends Component {
   };
 
   render() {
-    return (
-      <div id={this.name2}>
-        123
-        <Todo onClick={this.handleClick} text={a} name={this.name2} />
-        <Todo2 age={10} name={this.name4} />
-      </div>
-    );
+    return <div id={this.name2}>123</div>;
   }
 }
