+++ extensions/iceworks-refactor/src/test/suite/refactor.test.ts
@@ -1,6 +1,5 @@
 import * as assert from 'assert';
 import * as fs from 'fs';
-import * as vscode from 'vscode';
 import * as path from 'path';
 import { removeComponentCode } from '../../refactor';
 import prettierFormat from '../../utils/prettierFormat';
@@ -9,16 +8,24 @@ const testPath = path.resolve(__dirname, '../../../', 'src/test');
 const examplesPath = path.resolve(testPath, 'examples');
 const expectsPath = path.resolve(testPath, 'expects');
 
-suite('Extension Test Suite', () => {
-  vscode.window.showInformationMessage('Start all tests.');
-
-  test('Remove component from function components', (done) => {
+suite('Refactor: remove component in page', () => {
+  test('Remove component in function component', (done) => {
     const componentPath = path.join(examplesPath, 'components', 'Detail', 'index.tsx');
     const pagePath = path.join(examplesPath, 'pages', 'DetailPage', 'index.tsx');
     const code = removeComponentCode(pagePath, componentPath);
-    const expectCode = fs.readFileSync(path.join(expectsPath, 'DetailPage.tsx'), { encoding: 'utf-8' });
+    const expectCode = prettierFormat(fs.readFileSync(path.join(expectsPath, 'DetailPage.tsx'), { encoding: 'utf-8' }));
 
-    assert.strictEqual(code, prettierFormat(expectCode));
+    assert.strictEqual(code, expectCode);
+    done();
+  });
+
+  test('Remove component in class component', (done) => {
+    const componentPath = path.join(examplesPath, 'components', 'Todo', 'index.tsx');
+    const pagePath = path.join(examplesPath, 'pages', 'TodoPage', 'index.tsx');
+    const code = removeComponentCode(pagePath, componentPath);
+    const expectCode = prettierFormat(fs.readFileSync(path.join(expectsPath, 'TodoPage.tsx'), { encoding: 'utf-8' }));
+
+    assert.strictEqual(code, expectCode);
     done();
   });
 });
